Safari-611.1.10/LayoutTests/editing/input/paste-text-ending-with-interchange-newline.html

<!DOCTYPE html>
<html>
<body>
<textarea>abc
def
</textarea>
<input type="text">
<script src="../../resources/dump-as-markup.js"></script>
<script>

if (window.testRunner)
    testRunner.dumpAsText();

document.querySelector('textarea').focus();
document.execCommand('SelectAll', false, null);
document.execCommand('Copy', false, null);

var input = document.querySelector('input');
input.focus();
document.execCommand('Paste', false, null);

Markup.description("This tests pasting a text with an interchange new line at the end. WebKit shouldn't insert a new line.");
Markup.dump(input);

</script>
</body>
</html>
