safari-600.7-branch/LayoutTests/editing/input/paste-text-ending-with-interchange-newline-expected.txt

This tests pasting a text with an interchange new line at the end. WebKit shouldn't insert a new line.
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "abc def"
